Address review feedback on affinity group resources
Split CRUD out of resource.go into create, read, update, delete and
import files in both affinity group packages, matching the layout used
elsewhere in the provider. The move is behaviour-neutral: the unit tests
and the full acceptance suite pass unchanged.

Deprecate the cluster affinity group description attribute rather than
removing it. It is not backed by the API and has no effect, but removing
a shipped attribute breaks existing configurations, so it now carries a
deprecation message and will go in a future release.

Type server ids as int64 rather than narrowing them to int32. Compute
server ids are 64-bit, and the conversion was both unnecessary and
flagged as an overflow risk. The SDK change comes from the companion
spec fix.

Retarget the compute server managed override, which pointed at an
attribute that does not exist and so silently did nothing, and add the
missing copyright headers.

+++ internal/sdk/oapigen/model_save_cluster_affinity_group_request_affinity_group.go
@@ -28,7 +28,7 @@ type SaveClusterAffinityGroupRequestAffinityGroup struct {
 	Active *bool                                             `json:"active,omitempty"`
 	Pool   *SaveClusterAffinityGroupRequestAffinityGroupPool `json:"pool,omitempty"`
 	// List of Server IDs to include in the Affinity Group
-	Servers []int32 `json:"servers,omitempty"`
+	Servers []int64 `json:"servers,omitempty"`
 	// Visibility - Set to public to allow all tenants
 	Visibility *string `json:"visibility,omitempty"`
 	// Array of tenant account ids that are allowed access
